+++ examples/crd/.gitignore
@@ -1,0 +1,5 @@
+*.d.ts
+*.js
+node_modules
+dist/
+imports/
