+++ examples/crd/README.md
@@ -1,0 +1,6 @@
+# Custom Resource Definition Import Example
+
+This example application defines a Kubernetes app that uses a bunch of
+CustomResourceDefinitions (CRDs).
+
+The main thing to notice is the `imports` list in the `cdk8s.yaml` file.
