+++ examples/crd/cdk8s.yaml
@@ -1,0 +1,6 @@
+app: node index.js
+language: typescript
+imports:
+  - k8s@1.17.0
+  - https://raw.githubusercontent.com/jenkinsci/kubernetes-operator/master/deploy/crds/jenkins.io_jenkins_crd.yaml
+  - mattermost_crd.yaml
