+++ examples/crd/index.test.ts
@@ -1,0 +1,8 @@
+import { HelloKube } from './index';
+import { Testing } from 'cdk8s';
+
+test('snapshot', () => {
+  const app = Testing.app();
+  const chart = new HelloKube(app, 'hello');
+  expect(chart.toJson()).toMatchSnapshot();
+});
