+++ packages/cdk8s-cli/lib/import/crd.ts
@@ -1,0 +1,113 @@
+import { TypeGenerator } from './type-generator';
+import { ImportBase } from './base';
+import { CodeMaker } from 'codemaker';
+import { httpsGet } from '../util';
+import * as yaml from 'yaml';
+import * as fs from 'fs-extra';
+import * as path from 'path';
+
+interface CustomResourceDefinitionManifest {
+  apiVersion?: string;
+  kind?: string;
+  metadata?: {
+    name?: string;
+  };
+  spec?: {
+    group: string;
+    names: {
+      kind: string;
+    };
+    versions?: Array<{ name: string; schema?: { openAPIV3Schema?: any } }>;
+    validation?: { openAPIV3Schema?: any };
+  };
+}
+
+export class ImportCustomResourceDefinition extends ImportBase {
+  public static async match(source: string): Promise<undefined | CustomResourceDefinitionManifest> {
+    let manifest;
+    if (source.startsWith('https://')) {
+      manifest = yaml.parse(await httpsGet(source));
+    } else if (path.extname(source) === '.yaml' || path.extname(source) === '.yml' || path.extname(source) === '.json') {
+      if (!(await fs.pathExists(source))) {
+        throw new Error(`can't find file ${source}`);
+      }
+
+      manifest = yaml.parse(await fs.readFile(source, 'utf-8'));
+    }
+
+    if (!manifest) {
+      return undefined;
+    }
+
+    return manifest;
+  }
+
+  private readonly schema: any;
+  private readonly group: string;
+  private readonly version: string;
+  private readonly kind: string;
+  private readonly fqn: string;
+  
+  constructor(manifest: CustomResourceDefinitionManifest) {
+    super();
+
+    assert(manifest.apiVersion === 'apiextensions.k8s.io/v1beta1', '"apiVersion" must be "apiextensions.k8s.io/v1beta1"');
+    assert(manifest.kind === 'CustomResourceDefinition', '"kind" must be "CustomResourceDefinition"');
+
+    const spec = manifest.spec;
+    if (!spec) {
+      throw new Error(`manifest does not have a "spec" attribute`)
+    }
+  
+    const version = (spec.versions ?? [])[0];
+    if (!version) {
+      throw new Error(`unable to determine CRD version`);
+    }
+  
+    if (!manifest.metadata?.name) {
+      throw new Error(`"metadata.name" is required`);
+    }
+  
+    const schema = version.schema?.openAPIV3Schema ?? spec.validation?.openAPIV3Schema;
+  
+    if (!schema) {
+      throw new Error(`unable to determine schema for custom resource`);
+    }
+
+    this.schema = schema;
+    this.group = spec.group;
+    this.version = version.name;
+    this.kind = spec.names.kind;
+    this.fqn = this.kind;
+  }
+
+  public get moduleName() {
+    return this.kind.toLocaleLowerCase();
+  }
+
+  protected async generateTypeScript(code: CodeMaker) {
+
+
+    const types = new TypeGenerator();
+  
+    types.emitConstruct({
+      group: this.group,
+      version: this.version,
+      kind: this.kind,
+      fqn: this.fqn,
+      schema: this.schema
+    });
+
+    code.line(`// generated by cdk8s`);
+    code.line(`import { ApiObject } from 'cdk8s';`);
+    code.line(`import { Construct } from 'constructs';`);
+    code.line();
+    types.generate(code);
+  }
+}
+
+function assert(condition: boolean, message: string) {
+  if (!condition) {
+    throw new Error(`invalid CustomResourceDefinition manifest: ${message}`);
+  }
+}
